requirements.txt
jupyter
matplotlib
numpy
opencv-python
pillow
tensorflow-gpu==1.8.0
tqdm
